# src/swagger.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "API de Turismo",
        "summary": "Esta é uma API para uma aplicação web voltada para o Turismo. Através dessa API o front-end pode se comunicar com o back-end trafegando dados e permitindo a interação do usuário com a aplicação. Essa API permite a manipulação de recursos, como: Usuários, Estabelecimentos, Categorias de Estabelecimentos, Endereços, Favoritos, entre outros recursos da aplicação.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/users/{userId}": {
            "get": {
                "summary": "Listagem de Usuário pelo ID",
                "description": "Essa rota é responsável pelo listagem de um usuário específico através de seu ID na aplicação.",
                "tags": [
                    "Users"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/UserResponse"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Dados não encontrados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/establishments": {
            "post": {
                "summary": "Cadastro de Estabelecimento",
                "description": "Essa rota é responsável pelo cadastro de estabelecimento na aplicação.",
                "tags": [
                    "Establishments"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EstablishmentCreationInput"
                            },
                            "examples": {
                                "Establishment": {
                                    "value": {
                                        "name": "Minha Loja Exemplo",
                                        "cnpj": "12345678901234",
                                        "description": "Bem-vindos à nossa loja de exemplo!",
                                        "picture_url": "https://exemplo.com/imagem.jpg",
                                        "background_picture_url": "https://exemplo.com/fundo.jpg",
                                        "category_id": "123",
                                        "address": {
                                            "city": "Cidade Exemplo",
                                            "street": "Rua Exemplo",
                                            "number": "123",
                                            "neighborhood": "Bairro Exemplo",
                                            "complement": "Andar 2, Loja 3",
                                            "zip_code": "12345-678",
                                            "state": "Estado Exemplo",
                                            "country": "País Exemplo"
                                        },
                                        "workingTime": {
                                            "open_on_sunday": true,
                                            "open_on_monday": true,
                                            "open_on_tuesday": true,
                                            "open_on_wednesday": true,
                                            "open_on_thursday": true,
                                            "open_on_friday": true,
                                            "open_on_saturday": true,
                                            "opening_time": "09:00",
                                            "closing_time": "18:00"
                                        },
                                        "contacts": {
                                            "emails": [
                                                "email1@gmail.com",
                                                "email2@gmail.com",
                                                "email3@gmail.com"
                                            ],
                                            "phone_numbers": [
                                                "15999999999",
                                                "158888888888",
                                                "15777777777"
                                            ]
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Criado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/EstablishmentResponse"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Estabelecimento já existe",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "summary": "Listagem de Estabelecimentos",
                "description": "Essa rota é responsável pela listagem de estabelecimentos na aplicação.",
                "tags": [
                    "Establishments"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/EstablishmentsArrayResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Dados não encontrados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "summary": "Edição de Estabelecimento",
                "description": "Essa rota é responsável pela edição de estabelecimento na aplicação.",
                "tags": [
                    "Establishments"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EstablishmentUpdateInput"
                            },
                            "examples": {
                                "Establishment": {
                                    "value": {
                                        "id": "09fec66e-4201-4201-89f1-ac7e45482355",
                                        "name": "Minha Loja Exemplo",
                                        "cnpj": "12345678901234",
                                        "description": "Bem-vindos à nossa loja de exemplo!",
                                        "picture_url": "https://exemplo.com/imagem.jpg",
                                        "background_picture_url": "https://exemplo.com/fundo.jpg",
                                        "category_id": "123",
                                        "address": {
                                            "id": "3ea5aa0d-e58f-41f7-a8c2-b5f18501ae90",
                                            "city": "Cidade Exemplo",
                                            "street": "Rua Exemplo",
                                            "number": "123",
                                            "neighborhood": "Bairro Exemplo",
                                            "complement": "Andar 2, Loja 3",
                                            "zip_code": "12345-678",
                                            "state": "Estado Exemplo",
                                            "country": "País Exemplo"
                                        },
                                        "workingTime": {
                                            "id": "42e206d8-7bbe-416b-84d8-6b1123a6e37e",
                                            "open_on_sunday": true,
                                            "open_on_monday": true,
                                            "open_on_tuesday": true,
                                            "open_on_wednesday": true,
                                            "open_on_thursday": true,
                                            "open_on_friday": true,
                                            "open_on_saturday": true,
                                            "opening_time": "09:00",
                                            "closing_time": "18:00"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/EstablishmentResponse"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Dados não encontrados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "Deleção de Estabelecimento",
                "description": "Essa rota é responsável pela deleção de estabelecimento.",
                "tags": [
                    "Establishments"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "id": {
                                        "type": "string"
                                    }
                                },
                                "required": [
                                    "id"
                                ]
                            },
                            "examples": {
                                "Establishment": {
                                    "value": {
                                        "id": "67e60662-78e0-4260-8592-a0796da8e9dc"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/EstablishmentResponse"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Dados não encontrados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/establishments/{establishmentId}": {
            "get": {
                "summary": "Listagem de Estabelecimento pelo ID",
                "description": "Essa rota é responsável pelo listagem de um estabelecimento específico através de seu ID na aplicação.",
                "tags": [
                    "Establishments"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/EstablishmentResponse"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Dados não encontrados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/establishmentContacts/establishments/:id": {
            "get": {
                "summary": "Listagem de Contatos de um Estabelecimento pelo ID",
                "description": "Essa rota é responsável pelo listagem de contatos de um estabelecimento através de seu ID na aplicação.",
                "tags": [
                    "Establishment Contacts"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/EstablishmentContactArrayResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Dados não encontrados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/favoriteEstablishments/:id": {
            "get": {
                "summary": "Listagem de Favorito pelo ID",
                "description": "Essa rota é responsável pelo listagem de um favorito através de seu ID na aplicação.",
                "tags": [
                    "Favorite Establishments"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/FavoriteEstablishmentResponse"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Dados não encontrados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/favoriteEstablishments/users/:userId": {
            "get": {
                "summary": "Listagem de Favoritos de um Usuário pelo seu ID",
                "description": "Essa rota é responsável pelo listagem de favoritos de um usuário através de seu ID na aplicação.",
                "tags": [
                    "Favorite Establishments"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/FavoriteEstablishmentArrayResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Dados não encontrados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UserResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "picture_url": {
                        "type": "string"
                    },
                    "birthdate": {
                        "type": "string"
                    },
                    "gender": {
                        "type": "string"
                    },
                    "cpf": {
                        "type": "string"
                    },
                    "phone_number": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                }
            },
...
            "ErrorResponse": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string"
                    }
                }
            },
...
            "EstablishmentResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "cnpj": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "picture_url": {
                        "type": "string"
                    },
                    "background_picture_url": {
                        "type": "string"
                    },
                    "category_id": {
                        "type": "string"
                    },
                    "user_id": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                }
            },
            "EstablishmentsArrayResponse": {
                "type": "object",
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/EstablishmentResponse"
                        }
                    }
                }
            },
...
            "EstablishmentContactArrayResponse": {
                "type": "object",
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/EstablishmentContactResponse"
                        }
                    }
                }
            },
...
            "FavoriteEstablishmentResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "user_id": {
                        "type": "string"
                    },
                    "establishment_id": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                }
            },
            "FavoriteEstablishmentArrayResponse": {
                "type": "object",
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/FavoriteEstablishmentResponse"
                        }
                    }
                }
            }
        }
    },
...
}